add 'go install -a ./...' to 'make dependencies'
After upgrading a go version, all of the libraries need to be rebuilt or
the dependency upgrades can fail. Having 'go install -a ./...' be run
before the dependencies are fetched will make sure that users don't see
an unexpected error message.

This helps Sia to increasingly be a platform that 'just works', no
special knowledge needed to make sure everything builds correctly. We
still have a ways to go, but every step helps.